add services tests

+++ src/main/java/ru/caloriemanager/service/MealService.java
@@ -22,34 +22,42 @@ public class MealService {
     private MealRepository repository;
 
     public Meal create(Meal meal, int userId) {
+        if (meal == null || userId < 1) throw new IllegalArgumentException("not valid arguments");
         checkNew(meal);
         return repository.save(meal, userId);
     }
 
     public void update(Meal meal, int userId) {
+        if (meal == null || userId < 1) throw new IllegalArgumentException("not valid arguments");
         checkNotFoundWithId(repository.save(meal, userId), meal.getId());
     }
 
     public boolean delete(int id, int userId) {
+        if (id < 1 || userId < 1) throw new IllegalArgumentException("not valid arguments");
         checkNotFoundWithId(repository.get(id, userId), id);
         return repository.delete(id, userId);
     }
 
     public Meal get(int id, int userId) {
+        if (id < 1 || userId < 1) throw new IllegalArgumentException("not valid arguments");
         return checkNotFoundWithId(repository.get(id, userId), id);
     }
 
     public List<Meal> getAll(int userId) {
+        if (userId < 1) throw new IllegalArgumentException("not valid arguments");
         return repository.getAll(userId);
     }
 
     public List<Meal> getBetweenDates(@Nullable LocalDate startDate, @Nullable LocalDate endDate, int userId) {
+        if (userId < 1) throw new IllegalArgumentException("not valid arguments");
         return repository.getBetween(
                 DateTimeUtil.createDateTime(startDate, LocalDate.MIN, LocalTime.MIN),
                 DateTimeUtil.createDateTime(endDate, LocalDate.MAX, LocalTime.MAX), userId);
     }
 
     public List<Meal> getBetweenDatesTimes(LocalDateTime startDateTime, LocalDateTime endDateTime, int userId) {
+        if (userId < 1) throw new IllegalArgumentException("not valid arguments");
+        if (startDateTime == null || endDateTime == null) throw new IllegalArgumentException("not valid arguments");
         return repository.getBetween(startDateTime, endDateTime, userId);
     }
 }
